fix: 프로필 조회 모달을 어디서나 띄울 수 있게 함
